backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
sqlparse==0.5.3
pillow==11.1.0
celery==5.4.0
redis==5.2.1